注文管理と削除確認画面

+++ src/main/java/com/ikeda/repository/DvdItemRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.ikeda.entity.DvdItem;
 
-public interface DvdItemRepository extends JpaRepository<DvdItem, Integer> {
+public interface DvdItemRepository extends JpaRepository<DvdItem, Long> {
 
 }
